ContextAPI를 이용한 전역에서 상태 접근 및 관리하기
