+++ src/components/Todo/TodoItem.js
@@ -1,10 +1,14 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSquare } from '@fortawesome/free-regular-svg-icons';
 import { faCheckSquare, faTimes } from '@fortawesome/free-solid-svg-icons';
+import { TodoDispatch } from './Todo';
+import Store from '../../Store';
 
 /* To do 항목 */
-function TodoItem({ todo, onToggleCheckbox, onClickRemoveButton }) {
+function TodoItem({ todo }) {
+    const dispatch = useContext(TodoDispatch);
+
     return (
         <li className="todo-item">
             <input
@@ -17,13 +21,25 @@ function TodoItem({ todo, onToggleCheckbox, onClickRemoveButton }) {
                 <FontAwesomeIcon
                     className="todo-checkbox"
                     icon={faCheckSquare}
-                    onClick={() => onToggleCheckbox(todo.id)}
+                    onClick={() => {
+                        dispatch({
+                            type: 'TOGGLE_ITEM',
+                            id: todo.id,
+                        });
+                        Store.updateTodoList(todo.id);
+                    }}
                 />
             ) : (
                 <FontAwesomeIcon
                     className="todo-checkbox"
                     icon={faSquare}
-                    onClick={() => onToggleCheckbox(todo.id)}
+                    onClick={() => {
+                        dispatch({
+                            type: 'TOGGLE_ITEM',
+                            id: todo.id,
+                        });
+                        Store.updateTodoList(todo.id);
+                    }}
                 />
             )}
             <span className={todo.done ? 'todo-item-done' : ''}>
@@ -32,7 +48,13 @@ function TodoItem({ todo, onToggleCheckbox, onClickRemoveButton }) {
             <FontAwesomeIcon
                 className="todo-remove"
                 icon={faTimes}
-                onClick={() => onClickRemoveButton(todo.id)}
+                onClick={() => {
+                    dispatch({
+                        type: 'REMOVE_ITEM',
+                        id: todo.id,
+                    });
+                    Store.removeTodoItem(todo.id);
+                }}
             />
         </li>
     );
